Enhance complete_ledgers
Add ability to easily query if a ledger is in the server’s complete_ledgers set
