+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/client/server/ServerInfo.java
@@ -1,16 +1,24 @@
 package org.xrpl.xrpl4j.model.client.server;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.google.common.primitives.UnsignedInteger;
 import com.google.common.primitives.UnsignedLong;
 import org.immutables.value.Value;
+import org.immutables.value.Value.Derived;
 import org.immutables.value.Value.Immutable;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.xrpl.xrpl4j.model.immutables.FluentCompareTo;
 
 import java.time.ZonedDateTime;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
  * Maps the fields inside the "info" section of the "server_info" API call.
@@ -19,6 +27,8 @@ import java.util.Optional;
 @JsonSerialize(as = ImmutableServerInfo.class)
 @JsonDeserialize(as = ImmutableServerInfo.class)
 public interface ServerInfo {
+
+  Logger LOGGER = LoggerFactory.getLogger(ServerInfo.class);
 
   static ImmutableServerInfo.Builder builder() {
     return ImmutableServerInfo.builder();
@@ -44,21 +54,20 @@ public interface ServerInfo {
   String buildVersion();
 
   /**
-   * Information on the most recently closed ledger that has not been validated by consensus. If the most
-   * recently validated ledger is available, the response omits this field and includes {@link #validatedLedger()}
-   * instead.
+   * Information on the most recently closed ledger that has not been validated by consensus. If the most recently
+   * validated ledger is available, the response omits this field and includes {@link #validatedLedger()} instead.
    *
-   * @return An optionally-present {@link ServerInfoLedger} containing information about the server's view of the
-   *     most recently closed ledger.
+   * @return An optionally-present {@link ServerInfoLedger} containing information about the server's view of the most
+   *   recently closed ledger.
    */
   @JsonProperty("closed_ledger")
   Optional<ServerInfoLedger> closedLedger();
 
   /**
-   * Range expression indicating the sequence numbers of the ledger versions the local rippled has in its database.
-   * This may be a disjoint sequence such as {@code 24900901-24900984,24901116-24901158}.
-   * If the server does not have any complete ledgers (for example, it recently started syncing with the network),
-   * this will be an empty {@link String}.
+   * Range expression indicating the sequence numbers of the ledger versions the local rippled has in its database. This
+   * may be a disjoint sequence such as {@code 24900901-24900984,24901116-24901158}. If the server does not have any
+   * complete ledgers (for example, it recently started syncing with the network), this will be an empty {@link
+   * String}.
    *
    * @return A {@link String} representing a range of ledger sequences.
    */
@@ -66,8 +75,59 @@ public interface ServerInfo {
   String completeLedgers();
 
   /**
-   * On an admin request, returns the hostname of the server running the rippled instance;
-   * otherwise, returns a single RFC-1751 word based on the node public key.
+   * Determines if the supplied {@code ledgerIndex} exists on the rippled server by inspecting {@link
+   * #completeLedgers()}.
+   *
+   * @param ledgerIndex An {@link UnsignedLong} representing a particular ledger index.
+   *
+   * @return {@code true} if the rippled server includes {@code ledgerIndex} in its local database; {@code false}
+   *   otherwise.
+   */
+  @Derived
+  @JsonIgnore
+  default boolean isLedgerInCompleteLedgers(final UnsignedLong ledgerIndex) {
+    try {
+      return Optional.ofNullable(completeLedgers())
+        .filter($ -> !$.equalsIgnoreCase("empty"))
+        .map(completeLedgers -> {
+          // Split completeLedgers by comma...
+          final List<String> ranges = Stream.of(completeLedgers.split(",")).collect(Collectors.toList());
+          return ranges.stream()
+            .anyMatch(range -> {
+              // If ledgerIndex is in this range, return true; otherwise return false.
+
+              String[] parts = range.split("-");
+              if (parts.length == 1) {
+                return UnsignedLong.valueOf(parts[0]).equals(ledgerIndex);
+              }
+              if (parts.length == 2) {
+                UnsignedLong start = UnsignedLong.valueOf(parts[0]);
+                UnsignedLong end = UnsignedLong.valueOf(parts[1]);
+
+                if (
+                  FluentCompareTo.is(ledgerIndex).greaterThanEqualTo(start) &&
+                    FluentCompareTo.is(ledgerIndex).lessThanOrEqualTo(end)
+                ) {
+                  return true; // <-- ledgerIndex is in this range.
+                } else {
+                  return false; // <-- ledgerIndex not in this range (but try the other ranges).
+                }
+              } else {
+                throw new IllegalArgumentException("Range had too many dashes (should have had 0 or 1)");
+              }
+            });
+        })
+        .orElse(false);
+    } catch (Exception e) {
+      final String errorMessage = String.format("Unable to parse complete_ledgers value `%s`", completeLedgers());
+      LOGGER.warn(errorMessage, e);
+      return false;
+    }
+  }
+
+  /**
+   * On an admin request, returns the hostname of the server running the rippled instance; otherwise, returns a single
+   * RFC-1751 word based on the node public key.
    *
    * @return A {@link String} containing the host name of the serverk.
    */
@@ -75,8 +135,8 @@ public interface ServerInfo {
   String hostId();
 
   /**
-   * Amount of time spent waiting for I/O operations, in milliseconds. If this number is not very, very low,
-   * then the rippled server is probably having serious load issues.
+   * Amount of time spent waiting for I/O operations, in milliseconds. If this number is not very, very low, then the
+   * rippled server is probably having serious load issues.
    *
    * @return An {@link UnsignedLong} representing the I/O latency.
    */
@@ -84,9 +144,9 @@ public interface ServerInfo {
   UnsignedLong ioLatencyMs();
 
   /**
-   * The number of times (since starting up) that this server has had over 250 transactions waiting to be processed
-   * at once. A large number here may mean that your server is unable to handle the transaction load of the
-   * XRP Ledger network. For detailed recommendations of future-proof server specifications, see
+   * The number of times (since starting up) that this server has had over 250 transactions waiting to be processed at
+   * once. A large number here may mean that your server is unable to handle the transaction load of the XRP Ledger
+   * network. For detailed recommendations of future-proof server specifications, see
    * <a href="https://xrpl.org/capacity-planning.html">Capacity Planning</a>.
    *
    * @return A {@link String}.
@@ -112,11 +172,11 @@ public interface ServerInfo {
   Optional<ServerInfoLoad> load();
 
   /**
-   * The load-scaled open ledger transaction cost the server is currently enforcing, as a multiplier on the
-   * base transaction cost. For example, at 1000 load factor and a reference transaction cost of 10 drops of XRP,
-   * the load-scaled transaction cost is 10,000 drops (0.01 XRP). The load factor is determined by the highest
-   * of the individual server's load factor, the cluster's load factor, the open ledger cost and the overall
-   * network's load factor.
+   * The load-scaled open ledger transaction cost the server is currently enforcing, as a multiplier on the base
+   * transaction cost. For example, at 1000 load factor and a reference transaction cost of 10 drops of XRP, the
+   * load-scaled transaction cost is 10,000 drops (0.01 XRP). The load factor is determined by the highest of the
+   * individual server's load factor, the cluster's load factor, the open ledger cost and the overall network's load
+   * factor.
    *
    * @return An {@link UnsignedInteger} representing the load factor.
    */
@@ -132,8 +192,8 @@ public interface ServerInfo {
   Optional<UnsignedInteger> loadFactorLocal();
 
   /**
-   * Current multiplier to the transaction cost being used by the rest of the network
-   * (estimated from other servers' reported load values).
+   * Current multiplier to the transaction cost being used by the rest of the network (estimated from other servers'
+   * reported load values).
    *
    * @return An optionally-present {@link UnsignedInteger} representing the network load factor.
    */
@@ -157,8 +217,8 @@ public interface ServerInfo {
   Optional<UnsignedInteger> loadFactorFeeEscalation();
 
   /**
-   * The current multiplier to the transaction cost that a transaction must pay to get into the queue,
-   * if the queue is full.
+   * The current multiplier to the transaction cost that a transaction must pay to get into the queue, if the queue is
+   * full.
    *
    * @return An optionally-present {@link UnsignedInteger} representing the queue load factor.
    */
@@ -182,9 +242,9 @@ public interface ServerInfo {
 
   /**
    * Public key used to verify this server for peer-to-peer communications. This node key pair is automatically
-   * generated by the server the first time it starts up. (If deleted, the server can create a new pair of keys.)
-   * You can set a persistent value in the config file using the {@code [node_seed]} config option, which is useful
-   * for clustering.
+   * generated by the server the first time it starts up. (If deleted, the server can create a new pair of keys.) You
+   * can set a persistent value in the config file using the {@code [node_seed]} config option, which is useful for
+   * clustering.
    *
    * @return A {@link String} containing the node's public key.
    */
@@ -192,8 +252,8 @@ public interface ServerInfo {
   String publicKeyNode();
 
   /**
-   * (Admin only) Public key used by this node to sign ledger validations. This validation key pair is derived
-   * from the {@code [validator_token]} or {@code [validation_seed]} config field.
+   * (Admin only) Public key used by this node to sign ledger validations. This validation key pair is derived from the
+   * {@code [validator_token]} or {@code [validation_seed]} config field.
    *
    * @return A {@link String} containing the validator's public key.
    */
@@ -234,8 +294,8 @@ public interface ServerInfo {
   UnsignedLong upTime();
 
   /**
-   * Information about the most recent fully-validated ledger. If the most recent validated ledger is not available,
-   * the response omits this field and includes {@link #closedLedger()} instead.
+   * Information about the most recent fully-validated ledger. If the most recent validated ledger is not available, the
+   * response omits this field and includes {@link #closedLedger()} instead.
    *
    * @return An optionally-present {@link ServerInfoLedger} representing the latest validated ledger.
    */
@@ -243,8 +303,8 @@ public interface ServerInfo {
   Optional<ServerInfoLedger> validatedLedger();
 
   /**
-   * Minimum number of trusted validations required to validate a ledger version. Some circumstances may cause
-   * the server to require more validations.
+   * Minimum number of trusted validations required to validate a ledger version. Some circumstances may cause the
+   * server to require more validations.
    *
    * @return An {@link UnsignedInteger} representing the quorum.
    */
@@ -252,13 +312,12 @@ public interface ServerInfo {
   UnsignedInteger validationQuorum();
 
   /**
-   * (Admin only) Either the human readable time, in UTC, when the current validator list will expire,
-   * the string {@code "unknown"} if the server has yet to load a published validator list or the string
-   * {@code "never"} if the server uses a static validator list.
+   * (Admin only) Either the human readable time, in UTC, when the current validator list will expire, the string {@code
+   * "unknown"} if the server has yet to load a published validator list or the string {@code "never"} if the server
+   * uses a static validator list.
    *
    * @return An optionally-present {@link String} containing the validator expiration list.
    */
   @JsonProperty("validator_list_expires")
   Optional<String> validatorListExpires();
-
 }
